Move owner/repo/... parsing to GitForge type
